+++ schemas/contactSchemas.js
@@ -1,0 +1,19 @@
+const Joi = require("joi");
+
+const phoneRegexp = /^\(\d{3}\) \d{3}-\d{4}$/;
+
+const contactAddSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string().regex(phoneRegexp).required(),
+  favorite: Joi.boolean(),
+});
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+module.exports = {
+  contactAddSchema,
+  updateFavoriteSchema,
+};
